# docs/swagger.yaml
definitions:
...
  friend.friendInfo:
    properties:
      birth:
        type: string
      comment:
        type: string
      email:
        type: string
      ex:
        type: string
      gender:
        type: integer
      icon:
        type: string
      isInBlackList:
        type: integer
      mobile:
        type: string
      name:
        type: string
      uid:
        type: string
    type: object
...
  friend.paramsGetFriendLIst:
    properties:
      operationID:
        type: string
    required:
    - operationID
    type: object
...
  friend.paramsSearchFriend:
    properties:
      operationID:
        type: string
      ownerUid:
        type: string
      uid:
        type: string
    required:
    - operationID
    - uid
    type: object
  friend.paramsSetFriendComment:
    properties:
      comment:
        type: string
      operationID:
        type: string
      uid:
        type: string
    required:
    - operationID
    - uid
    type: object
...
  group.paramsCreateGroup:
    properties:
      ex:
        type: string
      faceUrl:
        type: string
      groupName:
        type: string
      introduction:
        type: string
      memberList:
        properties:
          setRole:
            type: string
          uid:
            type: string
        type: object
      notification:
        type: string
      operationID:
        type: string
    required:
    - operationID
    type: object
...
  group.paramsGroupApplicationList:
    properties:
      operationID:
        type: string
    required:
    - operationID
    type: object
  group.paramsGroupApplicationListRet:
    properties:
      createTime:
        type: integer
      flag:
        type: integer
      fromUserFaceURL:
        type: string
      fromUserID:
        type: string
      fromUserNickName:
        type: string
      groupID:
        type: string
      handleResult:
        type: integer
      handleStatus:
        type: integer
      handledMsg:
        type: string
      handledUser:
        type: string
      id:
        type: string
      reqMsg:
        type: string
      toUserFaceURL:
        type: string
      toUserID:
        type: string
      toUserNickName:
        type: string
      type:
        type: integer
    type: object
...
  group.resultGroupApplication:
    properties:
      count:
        type: integer
      user:
        items:
          $ref: '#/definitions/group.paramsGroupApplicationListRet'
        type: array
    type: object
...
  user.result:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errMsg:
        example: error
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /friend/get_friend_list:
    post:
      consumes:
      - application/json
      description: get friend apply list
      parameters:
      - description: get friend apply list
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsGetFriendLIst'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/friend.friendInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
  /friend/get_friends_info:
    post:
      consumes:
      - application/json
      description: get friend info
      parameters:
      - description: search friend params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsSearchFriend'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
...
  /friend/remove_blacklist:
    post:
      consumes:
      - application/json
      description: remove black list
      parameters:
      - description: remove black list params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsSearchFriend'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
  /friend/set_friend_comment:
    post:
      consumes:
      - application/json
      description: set friend comment
      parameters:
      - description: set friend comment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsSetFriendComment'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
  /group/create_group:
    post:
      consumes:
      - application/json
      description: create group
      parameters:
      - description: create group params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.paramsCreateGroup'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - group
...
  /group/get_group_applicationList:
    post:
      consumes:
      - application/json
      description: get group application list
      parameters:
      - description: get group application list params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.paramsGroupApplicationList'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  $ref: '#/definitions/group.resultGroupApplication'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - group
...
swagger: "2.0"